Check both Core Storage headers when checking for FileVault.
Core Storage header is stored at beginning and end of partition and are
usually identical.  When FV2 is first enabled, only the header at the
end of the partition is present, so look for both.  When additional
checks on the header are added, this will also allow the use of the
other header if the first is corrupt.